postitdaddy/requirements.txt
asgiref==3.4.1
auditlog==0.0.5
Django==4.0
django-etc==1.3.1
django-hitcount==1.3.3
django-jsonfield==1.4.1
django-resized==0.3.11
django-taggit==2.0.0
django-tinymce==3.4.0
Pillow==8.4.0
psycopg2==2.9.2
python-dateutil==2.6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
